4769: Changes for code review. Move logic back into ProjectStep.
