+++ spec/factories/project_groups.rb
@@ -23,33 +23,7 @@ FactoryGirl.define do
   factory :project_group do
     association :project, factory: :loan
     association :agent, factory: :person
-    scheduled_date { Faker::Date.between(Date.civil(2004, 01, 01), Date.today) }
-    is_finalized [true, false].sample
-    step_type_value { ["step", "milestone"] }
     transient_division
     summary { Faker::Hipster.paragraph }
-    details { Faker::Hipster.paragraphs }
-
-    trait :completed do
-      completed_date { Faker::Date.between(scheduled_date, Date.today) }
-    end
-
-    trait :past do
-      scheduled_date { Faker::Date.backward }
-    end
-
-    trait :future do
-      scheduled_date { Faker::Date.forward }
-    end
-
-    trait :with_logs do
-      after(:create) do |step|
-        create_list(
-          :project_log,
-          num_logs = 2,
-          project_step: step
-        )
-      end
-    end
   end
 end
