--- spec/support/modules/timeline_entry.rb
@@ -1,29 +1,0 @@
-shared_examples_for 'timeline_entry' do |attribute_names|
-  it 'can not be unfinalized after 24 hours' do
-    step = create(:project_step, is_finalized: true, finalized_at: Time.now-25.hours)
-    step.is_finalized = false
-    expect(step).to be_invalid
-  end
-
-  it 'can be unfinalized within 24 hours' do
-    step = create(:project_step, is_finalized: true, finalized_at: Time.now-23.hours)
-    step.is_finalized = false
-    expect(step).to be_valid
-  end
-
-  it 'has finalized_at automatically assigned' do
-    step = create(:project_step, is_finalized: false)
-    expect(step.finalized_at).to be_nil
-    step.update(is_finalized: true)
-    expect(step.finalized_at).not_to be_nil
-  end
-
-  it 'has original_date automatically assigned' do
-    step = create(:project_step, scheduled_date: Date.today, is_finalized: true)
-    expect(step[:original_date]).to be_nil
-    step.update(scheduled_date: Date.today + 2.days)
-    # Beware, the 'orginal_date' method will automatically returned scheduled date even when the
-    # raw value is nil, so need to directly check the attribute
-    expect(step[:original_date]).not_to be_nil
-  end
-end
